core/vision_tools.py
import os
import base64
from io import BytesIO
from PIL import Image
from typing import Dict, Any, Optional

class VisionTools:
    def __init__(self, api_key: Optional[str] = None):
        self.api_key = api_key or os.getenv("VISION_API_KEY")
        if not self.api_key:
            raise ValueError("Vision API key not provided and not found in environment variables.")
        # Initialize your vision model client here, e.g., GPT-4o-mini client

    async def analyze_screenshot(self, screenshot_path: str) -> Dict[str, Any]:
        # Read the screenshot and convert to base64
        with open(screenshot_path, "rb") as f:
            image_data = f.read()
            base64_image = base64.b64encode(image_data).decode("utf-8")

        # Call your vision model API here
        # This is a placeholder for the actual API call
        print(f"Analyzing screenshot: {screenshot_path} with vision model...")
        
        # Example response structure
        return {
            "description": "This is a placeholder description of the screenshot.",
            "action_coordinates": {"x": 100, "y": 200}, # Example coordinates
            "confidence": 0.9
        }

    async def click_coordinates(self, page, x: int, y: int):
        await page.mouse.click(x, y)

# Example usage (for testing purposes)
if __name__ == "__main__":
    async def test_vision_tools():
        # This example requires a dummy screenshot file and a Playwright page instance
        # For a full test, you'd need to set up Playwright and take a real screenshot
        print("VisionTools example requires a real screenshot and Playwright setup.")
        print("Skipping direct execution. Please integrate with a browser instance for testing.")

    import asyncio
    asyncio.run(test_vision_tools())