requirements.txt
langgraph
playwright
beautifulsoup4
requests
Pillow